pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "opencv-python-headless>=4.8.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "xlsxwriter>=3.2.3",
    "requests>=2.32.3",
    "opencv-python>=4.11.0.86",
    "pytesseract>=0.3.13",
    "spacy>=3.8.5",
    "sendgrid>=6.11.0",
    "openai>=1.77.0",
    "nltk>=3.9.1",
    "scikit-learn>=1.6.1",
    "pymupdf>=1.25.5",
    "numpy>=2.2.5",
    "reportlab>=4.4.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "slack-sdk>=3.35.0",
    "flask-socketio>=5.5.1",
    "eventlet>=0.39.1",
]
